requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.7.4.post0=py36h9ed2024_2
astor=0.8.1=py36hecd8cb5_0
async-timeout=3.0.1=py36hecd8cb5_0
attrs=21.4.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
boto3=1.21.32=pyhd3eb1b0_0
botocore=1.24.32=pyhd3eb1b0_0
brotlipy=0.7.0=py36h9ed2024_1003
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.07.19=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py36hecd8cb5_0
cffi=1.14.6=py36h2125817_0
chardet=4.0.0=py36hecd8cb5_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
coverage=5.5=py36h9ed2024_2
cryptography=35.0.0=py36h2fd3fbb_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.23=py36h23ab428_0
dataclasses=0.8=pyh4f3eec9_6
freetype=2.11.0=hd8bbffd_0
gast=0.2.2=py36_0
gensim=4.0.1=py36h23ab428_0
google-api-core=1.25.1=pyhd3eb1b0_0
google-auth=1.33.0=pyhd3eb1b0_0
google-cloud-core=1.6.0=pyhd3eb1b0_0
google-cloud-storage=1.40.0=pyhd3eb1b0_0
google-crc32c=1.1.2=py36h9ed2024_0
google-pasta=0.2.0=pyhd3eb1b0_0
google-resumable-media=1.3.1=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py36hecd8cb5_0
grpcio=1.36.1=py36h97de6d8_1
h5py=2.10.0=py36h0601b69_1
hdf5=1.10.6=hdbbcd12_0
idna=3.3=pyhd3eb1b0_0
idna_ssl=1.1.0=py36hecd8cb5_0
importlib-metadata=4.8.1=py36hecd8cb5_0
intel-openmp=2022.0.0=hecd8cb5_3615
jmespath=0.10.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
keras=2.3.1=0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py36_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py36h23ab428_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcrc32c=1.1.1=hb1e8313_2
libcxx=12.0.0=h2f01273_0
libdeflate=1.8=h9ed2024_5
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.17.2=h2842e9f_1
libtiff=4.4.0=h2ef1027_0
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.2.2=hca72f7f_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=py36hecd8cb5_0
matplotlib=3.3.4=py36hecd8cb5_0
matplotlib-base=3.3.4=py36h8b3ea08_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.3.0=py36ha059aab_0
mkl_random=1.1.1=py36h959d312_0
multidict=5.1.0=py36h9ed2024_2
ncurses=6.3=hca72f7f_3
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
nltk=3.4.5=py36_0
numpy=1.19.2=py36h456fd55_0
numpy-base=1.19.2=py36hcfb5961_0
olefile=0.46=py36_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1q=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
pandas=1.1.5=py36hb2f4e1b_0
pillow=8.3.1=py36ha4cf6ea_0
pip=21.2.2=py36hecd8cb5_0
protobuf=3.17.2=py36h23ab428_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py36hecd8cb5_0
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.1=py3.6_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=py36h9ed2024_1
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.5.0=pyhd3eb1b0_0
scipy=1.5.2=py36h912ce22_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=58.0.4=py36hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.1.0=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.2=h707629a_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=mkl_py36ha38f243_0
tensorflow-base=2.0.0=mkl_py36h66b1bf0_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py36hecd8cb5_1
tk=8.6.12=h5d9f67b_0
torchvision=0.2.2=py_3
tornado=6.1=py36h9ed2024_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.12.1=py36h1de35cc_1
xz=5.2.5=hca72f7f_1
yaml=0.2.5=haf1e3a3_0
yarl=1.5.1=py36haf1e3a3_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
